+++ components/layout/Layout.tsx
@@ -1,8 +1,8 @@
 import React, {ReactNode} from 'react';
-import Sidebar from "./Sidebar"
-import Topbar from "./Topbar";
-import {useAuth} from "../lib/auth";
-import SignIn from "../pages/sign_in";
+import Sidebar from "../Sidebar"
+import Topbar from "../Topbar";
+import {useAuth} from "../../lib/auth";
+import SignIn from "../../pages/sign_in";
 import {useRouter} from 'next/router'
 
 type layoutProps = {
